predefined boxes for transformer model, there are many new predefined boxes so there are many variations how to draw up the transformer model on the diagram and then generate code from the diagram.
Also fixed an issue with the deleting a box also fixed an issue where too many wires caused the application to throw an error.

Also changed the functionality how the spiders are handled when generating code from the diagram so that they could work for my needs when drawing neural network diagrams.