added 2 more ways to create the transformer model on the diagram and compile it into code.
Also tried to make the canvas bigger but this does not work correctly yet but the scroll bars are there.

also you can create a sub diagram and the code generation works but if there is nested sub diagrams or more than one sub diagram the code generation would not work correctly yet.